+++ src/classes/LipSync.js
@@ -2,6 +2,7 @@ const
     Cheer = require('./Cheer'),
     fs = require('fs').promises,
     getJSON = require('../helpers/getJSON'),
+    {resolveJobExports} = require('../helpers/jobExports'),
     {ensureDir, ensureParentDir, readdirOrEmpty} = require('../helpers/dirs'),
     id3 = require('node-id3').Promise,
     filter = (fileType, file) => (file.indexOf('.') !== 0) && (file.slice(-(fileType.length + 1)) === `.${fileType}`),
@@ -14,7 +15,7 @@ const
             const
                 file = files[i];
 
-            if (fileType === 'mp3') { // check id3 for lyrics presence.
+            if (fileType === 'mp3') {
                 const
                     {synchronisedLyrics} = await id3.read(`${path}${file}`);
 
@@ -54,9 +55,13 @@ const
 
         async checkDifference () {
             const
-                {config} = this,
-                {files = {}, options = {}, output, src} = config,
-                {exportFormat = 'json', output: outputType = 'json'} = options,
+                {config, service} = this,
+                {files = {}, options = {}, output, src, exports: configExports} = config,
+                exportFormats = resolveJobExports({
+                    exports: configExports,
+                    options
+                }, service, {defaultFormats: ['json']}),
+                usesJsonAggregate = exportFormats.includes('json'),
                 file = 'mouthCues.json';
 
             if (output) {
@@ -64,47 +69,128 @@ const
             }
 
             const
-                raw = outputType === 'json' && exportFormat !== 'mp3' ? await getJSON(`${output}${file}`) ?? {} : null,
-                alreadyLipSynced = raw ? Object.keys(raw) : await combine(output, exportFormat),
-                check = (id) => {
-                    const
-                        index = alreadyLipSynced.indexOf(id);
-
-                    if (index >= 0) {
-                        alreadyLipSynced.splice(index, 1);
-                        return true;
-                    } else {
-                        return false;
+                raw = usesJsonAggregate ? await getJSON(`${output}${file}`) ?? {} : null,
+                formatLists = {},
+                loadFormatList = async (formatId) => {
+                    if (formatId === 'json') {
+                        return raw ? Object.keys(raw) : [];
                     }
-                },
-                mergeLipSync = async (newLipSync = false) => { // Must run _after_ all checks and will remove what's left.
-                    for (let i = 0; i < alreadyLipSynced.length; i++) {
-                        const
-                            key = alreadyLipSynced[i];
 
+                    formatLists[formatId] = formatLists[formatId] ?? await combine(output, formatId);
+                    return formatLists[formatId];
+                },
+                markDone = (id) => {
+                    if (raw) {
+                        delete raw[id];
+                    }
+
+                    exportFormats.forEach((formatId) => {
+                        if (formatId === 'json') {
+                            return;
+                        }
+
+                        const
+                            list = formatLists[formatId];
+
+                        if (!list) {
+                            return;
+                        }
+
+                        const
+                            index = list.indexOf(id);
+
+                        if (index >= 0) {
+                            list.splice(index, 1);
+                        }
+                    });
+                },
+                isCurrent = async (id) => {
+                    for (let i = 0; i < exportFormats.length; i++) {
+                        const
+                            formatId = exportFormats[i],
+                            list = await loadFormatList(formatId);
+
+                        if (list.indexOf(id) < 0) {
+                            return false;
+                        }
+                    }
+
+                    markDone(id);
+                    return true;
+                },
+                mergeLipSync = async (newLipSync = false) => {
+                    const
+                        staleIds = new Set();
+
+                    if (raw) {
+                        Object.keys(raw).forEach((key) => staleIds.add(key));
+                    }
+
+                    exportFormats.forEach((formatId) => {
+                        if (formatId === 'json') {
+                            return;
+                        }
+
+                        (formatLists[formatId] || []).forEach((key) => staleIds.add(key));
+                    });
+
+                    for (const key of staleIds) {
                         if (raw) {
                             delete raw[key];
-                        } else {
-                            await fs.rm(`${output}${key}.${exportFormat}`);
                         }
+
+                        for (let i = 0; i < exportFormats.length; i++) {
+                            const
+                                formatId = exportFormats[i];
+
+                            if (formatId !== 'json') {
+                                await fs.rm(`${output}${key}.${formatId}`, {force: true});
+                            }
+                        }
+
                         console.log(`Removed "${key}"`);
                     }
+
                     if (raw) {
                         await ensureParentDir(`${output}${file}`);
                         await fs.writeFile(`${output}${file}`, JSON.stringify(sortKeys({
                             ...raw,
-                            ...newLipSync ? await getJSON(`${output}${file}`) ?? {} : {} // get new version.
+                            ...newLipSync ? await getJSON(`${output}${file}`) ?? {} : {}
                         }), null, 4));
                     }
 
-                    return alreadyLipSynced.length;
+                    return staleIds.size;
                 },
-                list = (await readdirOrEmpty(src)).filter(filter.bind(null, 'mp3')).filter((file) => !check(file.substring(0, file.length - 4)));
+                list = [];
+
+            for (const fileName of (await readdirOrEmpty(src)).filter(filter.bind(null, 'mp3'))) {
+                const
+                    id = fileName.substring(0, fileName.length - 4);
+
+                if (!(await isCurrent(id))) {
+                    list.push(fileName);
+                }
+            }
+
+            if (exportFormats.length === 1) {
+                config.options = {
+                    ...options,
+                    exportFormat: exportFormats[0]
+                };
+                delete config.exports;
+            } else {
+                config.exports = exportFormats;
+                config.options = {
+                    ...options,
+                    exportFormat: undefined,
+                    f: undefined
+                };
+            }
 
             this.differenceOnly = true;
+            this.exportFormats = exportFormats;
 
             if (list.length === 0) {
-                // We'll still run the merge in case any have been removed.
                 if (await mergeLipSync()) {
                     throw Error('Old lip-sync removed, but no new ones required generation.');
                 } else {